fix warning

+++ signer/src/signer/rrset.c
@@ -748,7 +748,7 @@ rrset_recycle(rrset_type* rrset, signconf_type* sc, time_t signtime)
 
     /* 1. If the RRset has changed, drop all signatures */
     /* 2. If Refresh is disabled, drop all signatures */
-    if (rrset->needs_signing || refresh <= signtime) {
+    if (rrset->needs_signing || refresh <= (uint32_t) signtime) {
         ods_log_debug("[%s] drop signatures for RRset[%i]", rrset_str,
             rrset->rr_type);
         if (rrset->rrsigs) {
